TASK-38 [Refactor]: dev 브랜치의 데이터 로직으로 통일
